pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto_video_pipeline"
version = "0.1.0"
description = "Automated video assembly pipeline"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Brant-B" }]
license = { text = "MIT" }
dependencies = [
  "typer[all]>=0.12.0",
  "pydantic>=2.6",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1",
  "ruff>=0.3",
]

[project.scripts]
auto-video-pipeline = "auto_video_pipeline.cli:main"
